fbneo paprium and fixes
- fix: fbneo samples out of memory error
- fix: fbneo rotation
- fix: fbneo paprium
- fix: synopsis scrolling
- mame and fbneo manufacturer, year and system in the main screen

+++ src/beans/structures.h
@@ -46,7 +46,7 @@ class Executable{
 };
 
 struct GameTicks{
-    uint16_t ticks;
+    uint32_t ticks;
 };
 
 struct Dimension{
